some static casts for windows to fix r24717
TBR=bmeurer@chromium.org

BUG=

Review URL: https://codereview.chromium.org/660843003

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@24718 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/instruction.cc
@@ -330,10 +330,10 @@ static BasicBlock::RpoNumber GetLoopEndRpo(const BasicBlock* block) {
 
 
 InstructionBlock::InstructionBlock(Zone* zone, const BasicBlock* block)
-    : successors_(block->SuccessorCount(), BasicBlock::RpoNumber::Invalid(),
-                  zone),
-      predecessors_(block->PredecessorCount(), BasicBlock::RpoNumber::Invalid(),
-                    zone),
+    : successors_(static_cast<int>(block->SuccessorCount()),
+                  BasicBlock::RpoNumber::Invalid(), zone),
+      predecessors_(static_cast<int>(block->PredecessorCount()),
+                    BasicBlock::RpoNumber::Invalid(), zone),
       phis_(zone),
       rpo_number_(block->GetRpoNumber()),
       loop_header_(GetRpo(block->loop_header())),
